social-media-analysis/backend/requirements.txt
flask
flask-cors
astrapy
python-dotenv
requests